+++ tests/test_version.py
@@ -2,4 +2,4 @@ from fabra import __version__
 
 
 def test_version() -> None:
-    assert __version__ == "2.0.3"
+    assert __version__ == "2.0.4"